fix(input): fix circular dependencies
